dev/chatbot-slack/requirements.txt
slackclient>=2.5.0
slackeventsapi>=2.1.0